+++ backend/baton/src/main/java/touch/baton/domain/oauth/OauthInformation.java
@@ -9,6 +9,7 @@ import touch.baton.domain.member.vo.ImageUrl;
 import touch.baton.domain.member.vo.MemberName;
 import touch.baton.domain.member.vo.OauthId;
 import touch.baton.domain.member.vo.SocialId;
+import touch.baton.domain.oauth.token.SocialToken;
 
 import static lombok.AccessLevel.PROTECTED;
