+++ backend/baton/src/main/java/touch/baton/domain/oauth/repository/RefreshTokenRepository.java
@@ -1,7 +1,7 @@
 package touch.baton.domain.oauth.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import touch.baton.domain.oauth.RefreshToken;
+import touch.baton.domain.oauth.token.RefreshToken;
 
 public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
 }
