+++ backend/baton/src/main/java/touch/baton/domain/oauth/service/OauthService.java
@@ -4,18 +4,18 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import touch.baton.domain.member.Member;
 import touch.baton.domain.member.vo.Company;
-import touch.baton.domain.oauth.AccessToken;
 import touch.baton.domain.oauth.OauthInformation;
 import touch.baton.domain.oauth.OauthType;
-import touch.baton.domain.oauth.RefreshToken;
-import touch.baton.domain.oauth.Token;
-import touch.baton.domain.oauth.Tokens;
 import touch.baton.domain.oauth.authcode.AuthCodeRequestUrlProviderComposite;
 import touch.baton.domain.oauth.client.OauthInformationClientComposite;
 import touch.baton.domain.oauth.repository.OauthMemberRepository;
 import touch.baton.domain.oauth.repository.OauthRunnerRepository;
 import touch.baton.domain.oauth.repository.OauthSupporterRepository;
 import touch.baton.domain.oauth.repository.RefreshTokenRepository;
+import touch.baton.domain.oauth.token.AccessToken;
+import touch.baton.domain.oauth.token.RefreshToken;
+import touch.baton.domain.oauth.token.Token;
+import touch.baton.domain.oauth.token.Tokens;
 import touch.baton.domain.runner.Runner;
 import touch.baton.domain.supporter.Supporter;
 import touch.baton.domain.supporter.vo.ReviewCount;
