+++ backend/baton/src/main/java/touch/baton/domain/oauth/token/RefreshToken.java
@@ -1,4 +1,4 @@
-package touch.baton.domain.oauth;
+package touch.baton.domain.oauth.token;
 
 import jakarta.persistence.Embedded;
 import jakarta.persistence.Entity;
