+++ backend/baton/src/main/java/touch/baton/domain/oauth/token/Tokens.java
@@ -1,4 +1,4 @@
-package touch.baton.domain.oauth;
+package touch.baton.domain.oauth.token;
 
 public record Tokens(
         AccessToken accessToken,
