+++ backend/baton/src/test/java/touch/baton/config/infra/auth/oauth/client/MockOauthInformationClientCompositeConfig.java
@@ -7,8 +7,8 @@ import touch.baton.config.infra.auth.oauth.authcode.MockAuthCodes;
 import touch.baton.domain.member.Member;
 import touch.baton.domain.oauth.OauthInformation;
 import touch.baton.domain.oauth.OauthType;
-import touch.baton.domain.oauth.SocialToken;
 import touch.baton.domain.oauth.client.OauthInformationClientComposite;
+import touch.baton.domain.oauth.token.SocialToken;
 import touch.baton.fixture.domain.MemberFixture;
 
 import static org.mockito.ArgumentMatchers.any;
